test/documentation: move tests to own file
