+++ src/documentation/chunk.rs
@@ -496,443 +496,3 @@ impl<'a> fmt::Display for ChunkDisplay<'a> {
         write!(formatter, "{}{}{}", ctx1, highlight, ctx2)
     }
 }
-
-#[cfg(test)]
-mod test {
-    use super::literalset::tests::gen_literal_set;
-    use super::util::load_span_from;
-    use super::*;
-
-    #[test]
-    fn find_spans_emoji() {
-        const TEST: &str = r##"ab **🐡** xy"##;
-
-        let chunk = CheckableChunk::from_str(
-            TEST,
-            indexmap::indexmap! { 0..11 => Span {
-                start: LineColumn {
-                    line: 1usize,
-                    column: 4usize,
-                },
-                end: LineColumn {
-                    line: 1usize,
-                    column: 14usize,
-                },
-            }},
-            CommentVariant::CommonMark,
-        );
-
-        assert_eq!(chunk.find_spans(0..2).len(), 1);
-        assert_eq!(chunk.find_spans(5..6).len(), 1);
-        assert_eq!(chunk.find_spans(9..11).len(), 1);
-        assert_eq!(chunk.find_spans(9..20).len(), 1);
-    }
-
-    #[test]
-    fn find_spans_simple() {
-        let _ = env_logger::builder()
-            .is_test(true)
-            .filter(None, log::LevelFilter::Trace)
-            .try_init();
-
-        // generate  `///<space>...`
-        const SOURCE: &'static str = fluff_up!(["xyz"]);
-        let set = gen_literal_set(SOURCE);
-        let chunk = dbg!(CheckableChunk::from_literalset(set));
-
-        // range in `chunk.as_str()`
-        // " xyz"
-        const CHUNK_RANGE: Range = 1..4;
-
-        // "/// xyz"
-        //  0123456
-        const EXPECTED_SPAN: Span = Span {
-            start: LineColumn { line: 1, column: 4 },
-            end: LineColumn { line: 1, column: 6 },
-        };
-
-        let range2span = chunk.find_spans(CHUNK_RANGE.clone());
-        // test deals only with a single line, so we know it only is a single entry
-        assert_eq!(range2span.len(), 1);
-
-        // assure the range is correct given the chunk
-        assert_eq!("xyz", &chunk.as_str()[CHUNK_RANGE.clone()]);
-
-        let (range, span) = dbg!(range2span.iter().next().unwrap());
-        assert!(CHUNK_RANGE.contains(&(range.start)));
-        assert!(CHUNK_RANGE.contains(&(range.end - 1)));
-        assert_eq!(
-            load_span_from(SOURCE.as_bytes(), dbg!(*span)).expect("Span extraction must work"),
-            "xyz".to_owned()
-        );
-        assert_eq!(span, &EXPECTED_SPAN);
-    }
-
-    #[test]
-    fn find_spans_multiline() {
-        let _ = env_logger::builder()
-            .is_test(true)
-            .filter(None, log::LevelFilter::Trace)
-            .try_init();
-
-        const SOURCE: &'static str = fluff_up!(["xyz", "second", "third", "Converts a span to a range, where `self` is converted to a range reltive to the",
-             "passed span `scope`."] @ "       "
-        );
-        let set = gen_literal_set(SOURCE);
-        let chunk = dbg!(CheckableChunk::from_literalset(set));
-        const SPACES: usize = 7;
-        const TRIPLE_SLASH_SPACE: usize = 4;
-        const CHUNK_RANGES: &[Range] =
-            &[1..4, (4 + 1 + 1 + 6 + 1 + 1)..(4 + 1 + 1 + 6 + 1 + 1 + 5)];
-        const EXPECTED_SPANS: &[Span] = &[
-            Span {
-                start: LineColumn {
-                    line: 1,
-                    column: SPACES + TRIPLE_SLASH_SPACE + 0,
-                },
-                end: LineColumn {
-                    line: 1,
-                    column: SPACES + TRIPLE_SLASH_SPACE + 2,
-                },
-            },
-            Span {
-                start: LineColumn {
-                    line: 3,
-                    column: SPACES + TRIPLE_SLASH_SPACE + 0,
-                },
-                end: LineColumn {
-                    line: 3,
-                    column: SPACES + TRIPLE_SLASH_SPACE + 4,
-                },
-            },
-            Span {
-                start: LineColumn {
-                    line: 4,
-                    column: SPACES + TRIPLE_SLASH_SPACE + 0,
-                },
-                end: LineColumn {
-                    line: 4,
-                    column: SPACES + TRIPLE_SLASH_SPACE + 78,
-                },
-            },
-            Span {
-                start: LineColumn {
-                    line: 5,
-                    column: SPACES + TRIPLE_SLASH_SPACE + 0,
-                },
-                end: LineColumn {
-                    line: 5,
-                    column: SPACES + TRIPLE_SLASH_SPACE + 19,
-                },
-            },
-        ];
-        const EXPECTED_STR: &[&'static str] = &[
-            "xyz",
-            "third",
-            "Converts a span to a range, where `self` is converted to a range reltive to the",
-            "passed span `scope`.",
-        ];
-
-        for (query_range, expected_span, expected_str) in itertools::cons_tuples(
-            CHUNK_RANGES
-                .iter()
-                .zip(EXPECTED_SPANS.iter())
-                .zip(EXPECTED_STR.iter()),
-        ) {
-            let range2span = chunk.find_spans(query_range.clone());
-            // test deals only with a single line, so we know it only is a single entry
-            assert_eq!(range2span.len(), 1);
-            let (range, span) = dbg!(range2span.iter().next().unwrap());
-            assert!(query_range.contains(&(range.start)));
-            assert!(query_range.contains(&(range.end - 1)));
-            assert_eq!(
-                load_span_from(SOURCE.as_bytes(), *span).expect("Span extraction must work"),
-                expected_str.to_owned()
-            );
-            assert_eq!(span, expected_span);
-        }
-    }
-
-    #[test]
-    fn find_spans_chyrp() {
-        let _ = env_logger::builder()
-            .is_test(true)
-            .filter(None, log::LevelFilter::Trace)
-            .try_init();
-
-        const SOURCE: &'static str = chyrp_up!(["Amsel", "Wacholderdrossel", "Buchfink"]);
-        let set = gen_literal_set(SOURCE);
-        let chunk = dbg!(CheckableChunk::from_literalset(set));
-
-        const CHUNK_RANGES: &[Range] = &[0..(5 + 1 + 16 + 1 + 8)];
-        const EXPECTED_SPANS: &[Span] = &[Span {
-            start: LineColumn {
-                line: 1,
-                column: 0 + 9,
-            }, // prefix is #[doc=r#"
-            end: LineColumn { line: 3, column: 7 }, // suffix is pointeless
-        }];
-
-        assert_eq!(
-            dbg!(&EXPECTED_SPANS[0]
-                .to_content_range(&chunk)
-                .expect("Must be ok to extract span from chunk")),
-            dbg!(&CHUNK_RANGES[0])
-        );
-
-        const EXPECTED_STR: &[&'static str] = &[r#"Amsel
-Wacholderdrossel
-Buchfink"#];
-
-        assert_eq!(EXPECTED_STR[0], chunk.as_str());
-
-        for (query_range, expected_span, expected_str) in itertools::cons_tuples(
-            CHUNK_RANGES
-                .iter()
-                .zip(EXPECTED_SPANS.iter())
-                .zip(EXPECTED_STR.iter()),
-        ) {
-            let range2span = chunk.find_spans(query_range.clone());
-            // test deals only with a single line, so we know it only is a single entry
-            assert_eq!(range2span.len(), 1);
-            let (range, span) = dbg!(range2span.iter().next().unwrap());
-            assert!(query_range.contains(&(range.start)));
-            assert!(query_range.contains(&(range.end - 1)));
-            assert_eq!(
-                load_span_from(SOURCE.as_bytes(), *span).expect("Span extraction must work"),
-                expected_str.to_owned()
-            );
-            assert_eq!(span, expected_span);
-        }
-    }
-
-    #[test]
-    fn find_spans_inclusive() {
-        let _ = env_logger::builder()
-            .filter_level(log::LevelFilter::Trace)
-            .is_test(true)
-            .try_init();
-
-        const SOURCE: &'static str = fluff_up!(["Some random words"]);
-        let set = gen_literal_set(SOURCE);
-        let chunk = dbg!(CheckableChunk::from_literalset(set));
-        // a range inside the span
-        const CHUNK_RANGE: Range = 4..15;
-
-        const EXPECTED_SPAN: Span = Span {
-            start: LineColumn { line: 1, column: 3 },
-            end: LineColumn {
-                line: 1,
-                column: 20,
-            },
-        };
-
-        let mut range2span = chunk.find_covered_spans(CHUNK_RANGE.clone());
-
-        // assure the range is correct given the chunk
-        assert_eq!("e random wo", &chunk.as_str()[CHUNK_RANGE.clone()]);
-
-        let span = dbg!(range2span.next().unwrap());
-        assert_eq!(
-            load_span_from(SOURCE.as_bytes(), dbg!(*span)).expect("Span extraction must work"),
-            " Some random words".to_owned()
-        );
-        assert_eq!(span, &EXPECTED_SPAN);
-        // test deals only with a single line, so we know it only is a single entry
-        assert_eq!(range2span.count(), 0);
-    }
-
-    #[test]
-    fn find_spans_and_coverage_integrity() -> Result<()> {
-        let _ = env_logger::builder()
-            .filter_level(log::LevelFilter::Trace)
-            .is_test(true)
-            .try_init();
-
-        const SOURCE: &'static str = fluff_up!(
-            [
-                "ü×ä×ß",
-                "DEF_HIJ",
-                "🐔🌗🐢"
-            ] @ "       "
-        );
-        let set = dbg!(gen_literal_set(SOURCE));
-        let chunk = dbg!(CheckableChunk::from_literalset(set));
-        let s = chunk.as_str();
-        assert_eq!(s, " ü×ä×ß\n DEF_HIJ\n 🐔🌗🐢");
-
-        const SPACES: usize = 7;
-        const TRIPLE_SLASH_SPACE: usize = 4;
-
-        const CHUNK_RANGE_START: usize = 5 + 1 + 1; // space + content + newline
-        const CHUNK_RANGE_END: usize = CHUNK_RANGE_START + 8;
-        const CHUNK_RANGE: Range = CHUNK_RANGE_START..CHUNK_RANGE_END;
-
-        const EXPECTED_SPANS: &[Span] = &[Span {
-            start: LineColumn {
-                line: 2,
-                column: SPACES + TRIPLE_SLASH_SPACE - 1, // we want to include the whitespace here
-            },
-            end: LineColumn {
-                line: 2,
-                column: SPACES + TRIPLE_SLASH_SPACE - 1 + 7, // inclusive
-            },
-        }];
-        dbg!(crate::util::sub_char_range(s, CHUNK_RANGE.clone()));
-        let coverage = chunk.find_covered_spans(CHUNK_RANGE.clone());
-        let mapping = chunk.find_spans(CHUNK_RANGE.clone());
-
-        for (coverage_span, (find_range, find_span), expected) in
-            itertools::cons_tuples(coverage.zip(mapping.iter()).zip(EXPECTED_SPANS))
-        {
-            let cs = crate::util::load_span_from(SOURCE.as_bytes(), coverage_span.clone())?;
-            let fs = crate::util::load_span_from(SOURCE.as_bytes(), find_span.clone())?;
-            let fr = crate::util::sub_char_range(chunk.as_str(), find_range.clone());
-            let x = crate::util::load_span_from(SOURCE.as_bytes(), expected.clone())?;
-            log::trace!("[find]chunk[range]: {:?}", fr);
-            log::trace!("[find]excerpt(true): {:?}", fs);
-            log::trace!("[cove]excerpt(true): {:?}", cs);
-            log::trace!("expected: {:?}", x);
-            assert_eq!(coverage_span, expected);
-            assert_eq!(find_span, expected);
-            assert_eq!(fr, x);
-            assert_eq!(fs, x);
-            assert_eq!(cs, x);
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn find_coverage_multiline() {
-        let _ = env_logger::builder()
-            .filter_level(log::LevelFilter::Trace)
-            .is_test(true)
-            .try_init();
-
-        const SOURCE: &'static str = fluff_up!(
-            [
-                "xyz",
-                "second",
-                "third",
-                "Converts a span to a range, where `self` is converted to a range reltive to the",
-                "passed span `scope`."
-            ] @ "       "
-        );
-        let set = dbg!(gen_literal_set(SOURCE));
-        let chunk = dbg!(CheckableChunk::from_literalset(set));
-        const SPACES: usize = 7;
-        const TRIPLE_SLASH_SPACE: usize = 3;
-        const CHUNK_RANGE: Range = 7..22;
-        const EXPECTED_SPANS: &[Span] = &[
-            Span {
-                start: LineColumn {
-                    line: 2,
-                    column: SPACES + TRIPLE_SLASH_SPACE + 0,
-                },
-                end: LineColumn {
-                    line: 2,
-                    column: SPACES + TRIPLE_SLASH_SPACE + 6,
-                },
-            },
-            Span {
-                start: LineColumn {
-                    line: 3,
-                    column: SPACES + TRIPLE_SLASH_SPACE + 0,
-                },
-                end: LineColumn {
-                    line: 3,
-                    column: SPACES + TRIPLE_SLASH_SPACE + 5,
-                },
-            },
-            Span {
-                start: LineColumn {
-                    line: 4,
-                    column: SPACES + TRIPLE_SLASH_SPACE + 0,
-                },
-                end: LineColumn {
-                    line: 4,
-                    column: SPACES + TRIPLE_SLASH_SPACE + 79,
-                },
-            },
-        ];
-
-        let coverage = chunk.find_covered_spans(CHUNK_RANGE);
-
-        for (span, expected) in coverage.zip(EXPECTED_SPANS) {
-            assert_eq!(span, expected);
-        }
-    }
-
-    #[test]
-    fn find_line_lengths_tripple_slash() {
-        const SOURCE: &'static str = fluff_up!(
-            [
-                "xyz",
-                "second",
-                "third",
-                "Converts a span to a range, where `self` is converted to a range reltive to the",
-                "passed span `scope`."
-            ]
-            @ "       "
-        );
-
-        let _ = env_logger::builder()
-            .is_test(true)
-            .filter_level(log::LevelFilter::Trace)
-            .try_init();
-
-        let set = gen_literal_set(SOURCE);
-        let chunk = dbg!(CheckableChunk::from_literalset(set));
-
-        let lens = chunk
-            .extract_line_lengths()
-            .expect("Chunk in unit test must have lines.");
-
-        // the demo creates a single `struct X` entity with `n` comments before that.
-        let expected_line_count = SOURCE.lines().count() - 1;
-        assert_eq!(lens.len(), expected_line_count);
-
-        // XXX WRONG
-        for (len, line) in lens.iter().zip(SOURCE.lines()) {
-            assert_eq!(len, &line.len());
-        }
-    }
-
-    #[test]
-    #[ignore = "prefix and suffix are not part of the accounted `extract_lines_lengths"]
-    fn find_line_length_docmacro() {
-        const SOURCE: &'static str = chyrp_up!(
-            [
-                "xyz",
-                "second",
-                "third",
-                "Converts a span to a range, where `self` is converted to a range relative to the",
-                "passed span `scope`."
-            ]
-            @ "       "
-        );
-
-        let _ = env_logger::builder()
-            .is_test(true)
-            .filter_level(log::LevelFilter::Trace)
-            .try_init();
-
-        println!("{}", SOURCE);
-        let set = gen_literal_set(SOURCE);
-        let chunk = dbg!(CheckableChunk::from_literalset(set));
-
-        let lens = chunk
-            .extract_line_lengths()
-            .expect("Chunk in unit test must have lines.");
-
-        // the demo creates a single `struct X` entity with `n` comments before that.
-        let expected_line_count = SOURCE.lines().count() - 1;
-        assert_eq!(lens.len(), expected_line_count);
-
-        // XXX WRONG
-        for (len, line) in lens.iter().zip(SOURCE.lines()) {
-            dbg!(len, &line);
-            assert_eq!(*len, line.chars().count());
-        }
-    }
-}
