+++ src/documentation/literal.rs
@@ -463,479 +463,3 @@ impl<'a> fmt::Display for TrimmedLiteralDisplay<'a> {
     }
 }
 
-#[cfg(test)]
-pub(crate) mod tests {
-    use super::*;
-    use crate::util::load_span_from;
-    use crate::LineColumn;
-
-    pub(crate) fn annotated_literals(source: &str) -> Vec<TrimmedLiteral> {
-        let stream =
-            syn::parse_str::<proc_macro2::TokenStream>(source).expect("Must be valid rust");
-        stream
-            .into_iter()
-            .filter_map(|x| {
-                if let proc_macro2::TokenTree::Group(group) = x {
-                    Some(group.stream().into_iter())
-                } else {
-                    None
-                }
-            })
-            .flatten()
-            .filter_map(|x| {
-                if let proc_macro2::TokenTree::Literal(literal) = x {
-                    Some(literal)
-                } else {
-                    None
-                }
-            })
-            .map(|literal| {
-                TrimmedLiteral::try_from((source, literal))
-                    .expect("Literals must be convertable to trimmed literals")
-            })
-            .collect()
-    }
-
-    const PREFIX_RAW_LEN: usize = 3;
-    const SUFFIX_RAW_LEN: usize = 2;
-    const GAENSEFUESSCHEN: usize = 1;
-
-    #[derive(Clone, Debug)]
-    struct Triplet {
-        /// source content
-        source: &'static str,
-        /// expected doc comment content without modifications
-        extracted: &'static str,
-        /// expected doc comment content after applying trimming rules
-        trimmed: &'static str,
-        /// expected span as extracted by proc_macro2
-        extracted_span: Span,
-        /// trimmed span, so it is aligned with the proper doc comment
-        trimmed_span: Span,
-        /// expected variant
-        variant: CommentVariant,
-    }
-
-    fn comment_variant_span_range_validation(index: usize) {
-        let test_data: &[Triplet] = &[
-            // 0
-            Triplet {
-                source: r#"
-/// One Doc
-struct One;
-"#,
-                extracted: r#"" One Doc""#,
-                trimmed: " One Doc",
-                extracted_span: Span {
-                    start: LineColumn {
-                        line: 2_usize,
-                        column: 0,
-                    },
-                    end: LineColumn {
-                        line: 2_usize,
-                        column: 10_usize,
-                    },
-                },
-                trimmed_span: Span {
-                    start: LineColumn {
-                        line: 2_usize,
-                        column: 3_usize,
-                    },
-                    end: LineColumn {
-                        line: 2_usize,
-                        column: 10_usize,
-                    },
-                },
-                variant: CommentVariant::TripleSlash,
-            },
-            // 1
-            Triplet {
-                source: r##"
-    ///meanie
-struct Meanie;
-"##,
-                extracted: r#""meanie""#,
-                trimmed: "meanie",
-                extracted_span: Span {
-                    start: LineColumn {
-                        line: 2_usize,
-                        column: 0_usize + 7 - PREFIX_RAW_LEN,
-                    },
-                    end: LineColumn {
-                        line: 2_usize,
-                        column: 0_usize + 12 + SUFFIX_RAW_LEN,
-                    },
-                },
-                trimmed_span: Span {
-                    start: LineColumn {
-                        line: 2_usize,
-                        column: 0_usize + 7,
-                    },
-                    end: LineColumn {
-                        line: 2_usize,
-                        column: 0_usize + 12,
-                    },
-                },
-                variant: CommentVariant::TripleSlash,
-            },
-            // 2
-            Triplet {
-                source: r#"
-#[doc = "Two Doc"]
-struct Two;
-"#,
-                extracted: r#""Two Doc""#,
-                trimmed: "Two Doc",
-                extracted_span: Span {
-                    start: LineColumn {
-                        line: 2_usize,
-                        column: 0_usize + 10 - GAENSEFUESSCHEN,
-                    },
-                    end: LineColumn {
-                        line: 2__usize,
-                        column: 6__usize + 10 + GAENSEFUESSCHEN,
-                    },
-                },
-                trimmed_span: Span {
-                    start: LineColumn {
-                        line: 2__usize,
-                        column: 0__usize + 9,
-                    },
-                    end: LineColumn {
-                        line: 2__usize,
-                        column: 6__usize + 9,
-                    },
-                },
-                variant: CommentVariant::MacroDocEq("#[doc = ".to_string(), 0),
-            },
-            // 3
-            Triplet {
-                source: r##"
-    #[doc=r#"Three Doc"#]
-struct Three;
-"##,
-                extracted: r###"r#"Three Doc"#"###,
-                trimmed: "Three Doc",
-                extracted_span: Span {
-                    start: LineColumn {
-                        line: 2_usize,
-                        column: 4__usize + 11,
-                    },
-                    end: LineColumn {
-                        line: 2__usize,
-                        column: 13__usize + 11,
-                    },
-                },
-                trimmed_span: Span {
-                    start: LineColumn {
-                        line: 2__usize,
-                        column: 0__usize + 13,
-                    },
-                    end: LineColumn {
-                        line: 2__usize,
-                        column: 13__usize + 8,
-                    },
-                },
-                variant: CommentVariant::MacroDocEq("#[doc=".to_string(), 2),
-            },
-            // 4
-            Triplet {
-                source: r###"
-#[doc = r##"Four
-has
-multiple
-lines
-"##]
-struct Four;
-"###,
-                extracted: r###"r##"Four
-has
-multiple
-lines
-"##"###,
-                trimmed: r#"Four
-has
-multiple
-lines
-"#,
-                extracted_span: Span {
-                    start: LineColumn {
-                        line: 2__usize,
-                        column: 12__usize - 4,
-                    },
-                    end: LineColumn {
-                        line: 6__usize,
-                        column: 0__usize + 3,
-                    },
-                },
-                trimmed_span: Span {
-                    start: LineColumn {
-                        line: 2__usize,
-                        column: 12__usize,
-                    },
-                    end: LineColumn {
-                        line: 6__usize,
-                        column: 0__usize,
-                    },
-                },
-                variant: CommentVariant::MacroDocEq("#[doc = ".to_string(), 3),
-            },
-            // 5
-            Triplet {
-                source: r###"
-#[doc        ="XYZ"]
-struct Five;
-"###,
-                extracted: r#""XYZ""#,
-                trimmed: r#"XYZ"#,
-                extracted_span: Span {
-                    start: LineColumn {
-                        line: 2__usize,
-                        column: 15__usize - GAENSEFUESSCHEN,
-                    },
-                    end: LineColumn {
-                        line: 2__usize,
-                        column: 15__usize + 2 + GAENSEFUESSCHEN,
-                    },
-                },
-                trimmed_span: Span {
-                    start: LineColumn {
-                        line: 2__usize,
-                        column: 15_usize,
-                    },
-                    end: LineColumn {
-                        line: 2_usize,
-                        column: 15_usize + 2,
-                    },
-                },
-                variant: CommentVariant::MacroDocEq("#[doc        =".to_string(), 0),
-            },
-            // 6
-            Triplet {
-                source: r#"
-
-    /// if a layer is provided a identiacla "input" and "output", it will only be supplied an
-    fn compute_in_place(&self) -> bool {
-        false
-    }
-
-"#,
-                extracted: r#"" if a layer is provided a identiacla "input" and "output", it will only be supplied an""#,
-                trimmed: r#" if a layer is provided a identiacla "input" and "output", it will only be supplied an"#,
-                extracted_span: Span {
-                    start: LineColumn {
-                        line: 3_usize,
-                        column: 7_usize - GAENSEFUESSCHEN,
-                    },
-                    end: LineColumn {
-                        line: 2_usize,
-                        column: 92_usize + GAENSEFUESSCHEN,
-                    },
-                },
-                trimmed_span: Span {
-                    start: LineColumn {
-                        line: 3_usize,
-                        column: 7_usize,
-                    },
-                    end: LineColumn {
-                        line: 3_usize,
-                        column: 92_usize,
-                    },
-                },
-                variant: CommentVariant::TripleSlash,
-            },
-            // 7
-            Triplet {
-                source: r#"
-
-/// 🍉 ← αA<sup>OP</sup>x + βy
-fn unicode(&self) -> bool {
-    true
-}
-
-"#,
-                extracted: r#"" 🍉 ← αA<sup>OP</sup>x + βy""#,
-                trimmed: r#" 🍉 ← αA<sup>OP</sup>x + βy"#,
-                extracted_span: Span {
-                    start: LineColumn {
-                        line: 3_usize,
-                        column: 3_usize - GAENSEFUESSCHEN,
-                    },
-                    end: LineColumn {
-                        line: 2_usize,
-                        column: 28_usize + GAENSEFUESSCHEN,
-                    },
-                },
-                trimmed_span: Span {
-                    start: LineColumn {
-                        line: 3_usize,
-                        column: 3_usize,
-                    },
-                    end: LineColumn {
-                        line: 3_usize,
-                        column: 28_usize,
-                    },
-                },
-                variant: CommentVariant::TripleSlash,
-            },
-            // 8
-            Triplet {
-                source: r###"
-        #[doc = r##"Four
-        has
-
-        multiple
-        lines
-        "##]
-        struct Four;
-        "###,
-                extracted: r###"r##"Four
-        has
-
-        multiple
-        lines
-        "##"###,
-                trimmed: r#"Four
-        has
-
-        multiple
-        lines
-        "#,
-                extracted_span: Span {
-                    start: LineColumn {
-                        line: 2_usize,
-                        column: 12_usize - 4,
-                    },
-                    end: LineColumn {
-                        line: 6_usize,
-                        column: 0_usize + 3,
-                    },
-                },
-                trimmed_span: Span {
-                    start: LineColumn {
-                        line: 2_usize,
-                        column: 12_usize,
-                    },
-                    end: LineColumn {
-                        line: 6_usize,
-                        column: 0_usize,
-                    },
-                },
-                variant: CommentVariant::MacroDocEq("#[ doc = ".to_string(), 3),
-            },
-        ];
-
-        let _ = env_logger::builder()
-            .filter(None, log::LevelFilter::Trace)
-            .is_test(true)
-            .try_init();
-
-        let triplet = test_data[index].clone();
-        let literals = annotated_literals(triplet.source);
-
-        assert_eq!(literals.len(), 1);
-
-        let literal = literals.first().expect("Must contain exactly one literal");
-
-        assert_eq!(literal.as_str(), triplet.trimmed);
-
-        // just for better visual errors
-        let excerpt = load_span_from(triplet.source.as_bytes(), literal.span()).unwrap();
-        let expected_excerpt =
-            load_span_from(triplet.source.as_bytes(), triplet.trimmed_span).unwrap();
-        assert_eq!(excerpt, expected_excerpt);
-
-        assert_eq!(literal.span(), triplet.trimmed_span);
-
-        assert_eq!(literal.variant(), triplet.variant);
-    }
-
-    #[test]
-    fn raw_variant_0_triple_slash() {
-        comment_variant_span_range_validation(0);
-    }
-
-    #[test]
-    fn raw_variant_1_spaces_triple_slash() {
-        comment_variant_span_range_validation(1);
-    }
-
-    #[test]
-    fn raw_variant_2_spaces_doc_eq_single_quote() {
-        comment_variant_span_range_validation(2);
-    }
-
-    #[test]
-    fn raw_variant_3_doc_eq_single_r_hash_quote() {
-        comment_variant_span_range_validation(3);
-    }
-
-    #[test]
-    fn raw_variant_4_doc_eq_multi() {
-        comment_variant_span_range_validation(4);
-    }
-
-    #[test]
-    fn raw_variant_5_doc_spaces_eq_single_quote() {
-        comment_variant_span_range_validation(5);
-    }
-
-    #[test]
-    fn raw_variant_6_quote_chars() {
-        comment_variant_span_range_validation(6);
-    }
-
-    #[test]
-    fn raw_variant_7_unicode_symbols() {
-        comment_variant_span_range_validation(7);
-    }
-
-    #[test]
-    fn variant_to_string() {
-        let variant = CommentVariant::MacroDocEq("#[ doc = ".to_string(), 0);
-        assert_eq!(variant.prefix_string(), r###"#[ doc = ""###);
-        let variant = CommentVariant::MacroDocEq("#[doc = ".to_string(), 1);
-        assert_eq!(variant.prefix_string(), r###"#[doc = r""###);
-        let variant = CommentVariant::MacroDocEq("#[ doc =".to_string(), 2);
-        assert_eq!(variant.prefix_string(), r###"#[ doc =r#""###);
-        let variant = CommentVariant::MacroDocEq("#[doc=".to_string(), 3);
-        assert_eq!(variant.prefix_string(), r###"#[doc=r##""###);
-    }
-
-    #[test]
-    fn variant_suffix_string() {
-        let variant = CommentVariant::MacroDocEq("#[ doc= ".to_string(), 0);
-        assert_eq!(variant.suffix_string(), r###""]"###);
-        let variant = CommentVariant::MacroDocEq("#[doc = ".to_string(), 1);
-        assert_eq!(variant.suffix_string(), r###""]"###);
-        let variant = CommentVariant::MacroDocEq("#[doc = ".to_string(), 2);
-        assert_eq!(variant.suffix_string(), r###""#]"###);
-        let variant = CommentVariant::MacroDocEq("#[ doc =".to_string(), 3);
-        assert_eq!(variant.suffix_string(), r###""##]"###);
-    }
-
-    #[test]
-    fn variant_consistency() {
-        let variants = vec![
-            CommentVariant::TripleSlash,
-            CommentVariant::DoubleSlashEM,
-            CommentVariant::CommonMark,
-            CommentVariant::MacroDocEq("#[ doc= ".to_string(), 0),
-            CommentVariant::MacroDocEq("#[doc = ".to_string(), 1),
-            CommentVariant::MacroDocEq("#[doc = ".to_string(), 2),
-            CommentVariant::MacroDocEq("#[ doc     =".to_string(), 3),
-        ];
-
-        for variant in variants {
-            let variant = dbg!(variant);
-            assert_eq!(
-                variant.prefix_string().chars().count(),
-                variant.prefix_len()
-            );
-            assert_eq!(
-                variant.suffix_string().chars().count(),
-                variant.suffix_len()
-            );
-        }
-    }
-}
