+++ src/documentation/markdown.rs
@@ -57,7 +57,7 @@ impl<'a> PlainOverlay<'a> {
     }
 
     /// Ranges are mapped `cmark reduced/plain -> raw`.
-    fn extract_plain_with_mapping(cmark: &str) -> (String, IndexMap<Range, Range>) {
+    pub(crate) fn extract_plain_with_mapping(cmark: &str) -> (String, IndexMap<Range, Range>) {
         let mut plain = String::with_capacity(cmark.len());
         let mut mapping = indexmap::IndexMap::with_capacity(128);
 
@@ -368,285 +368,5 @@ impl<'a> fmt::Debug for PlainOverlay<'a> {
         writeln!(formatter, "Commonmark:\n{}", coloured_md)?;
         writeln!(formatter, "Plain:\n{}", coloured_plain)?;
         Ok(())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn drill_span() {
-        const TEST: &str = r##"ab **🐡** xy"##;
-        let chunk = CheckableChunk::from_str(
-            TEST,
-            indexmap::indexmap! { 0..11 => Span {
-                start: LineColumn {
-                    line: 1usize,
-                    column: 4usize,
-                },
-                end: LineColumn {
-                    line: 1usize,
-                    column: 14usize,
-                },
-            }},
-            CommentVariant::CommonMark,
-        );
-
-        let plain = chunk.erase_cmark();
-        assert_eq!(plain.find_spans(0..2).len(), 1);
-        assert_eq!(plain.find_spans(3..4).len(), 1);
-        assert_eq!(plain.find_spans(5..7).len(), 1);
-        assert_eq!(plain.find_spans(5..12).len(), 1);
-        assert_eq!(plain.find_spans(9..20).len(), 0);
-    }
-
-    #[test]
-    fn reduction_complex() {
-        const MARKDOWN: &str = r##"# Title number 1
-
-## Title number 2
-
-```rust
-let x = 777;
-let y = 111;
-let z = x/y;
-assert_eq!(z,7);
-```
-
-### Title [number 3][ff]
-
-Some **extra** _formatting_ if __anticipated__ or _*not*_ or
-maybe not at all.
-
-
-Extra ~pagaph~ _paragraph_.
-
----
-
-And a line, or a **rule**.
-
-
-[ff]: https://docs.rs
-"##;
-
-        const PLAIN: &str = r##"Title number 1
-
-Title number 2
-
-Title number 3
-
-Some extra formatting if anticipated or not or
-maybe not at all.
-
-Extra ~pagaph~ paragraph.
-
-
-And a line, or a rule."##;
-        let (reduced, mapping) = PlainOverlay::extract_plain_with_mapping(MARKDOWN);
-
-        assert_eq!(dbg!(&reduced).as_str(), PLAIN);
-        assert_eq!(dbg!(&mapping).len(), 20);
-        for (reduced_range, cmark_range) in mapping.iter() {
-            assert_eq!(
-                reduced[reduced_range.clone()],
-                MARKDOWN[cmark_range.clone()]
-            );
-        }
-    }
-
-    #[test]
-    fn reduction_leading_space() {
-        const MARKDOWN: &str = r#"  Some __underlined__ **bold** text."#;
-        const PLAIN: &str = r#"Some underlined bold text."#;
-
-        let (reduced, mapping) = PlainOverlay::extract_plain_with_mapping(MARKDOWN);
-
-        assert_eq!(dbg!(&reduced).as_str(), PLAIN);
-        assert_eq!(dbg!(&mapping).len(), 5);
-        for (reduced_range, cmark_range) in mapping.iter() {
-            assert_eq!(
-                reduced[reduced_range.clone()].to_owned(),
-                MARKDOWN[cmark_range.clone()].to_owned()
-            );
-        }
-    }
-
-    #[test]
-    fn range_test() {
-        let mut x = IndexMap::<Range, Range>::new();
-        x.insert(0..2, 1..3);
-        x.insert(3..4, 7..8);
-        x.insert(5..12, 11..18);
-
-        let lookmeup = 6..8;
-
-        // TODO keep in sync with copy pasta source, extract a func for this
-        let plain_range = lookmeup;
-        let v: Vec<_> = x
-            .iter()
-            .filter(|(plain, _md)| plain.start <= plain_range.end && plain_range.start <= plain.end)
-            .fold(Vec::with_capacity(64), |mut acc, (plain, md)| {
-                // calculate the linear shift
-                let offset = dbg!(md.start - plain.start);
-                assert_eq!(md.end - plain.end, offset);
-                let extracted = Range {
-                    start: plain_range.start + offset,
-                    end: core::cmp::min(md.end, plain_range.end + offset),
-                };
-                acc.push(extracted);
-                acc
-            });
-        assert_eq!(v.first(), Some(&(12..14)));
-    }
-
-    fn cmark_reduction_test(
-        input: &'static str,
-        expected: &'static str,
-        expected_mapping_len: usize,
-    ) {
-        let (plain, mapping) = PlainOverlay::extract_plain_with_mapping(input);
-        assert_eq!(dbg!(&plain).as_str(), expected);
-        assert_eq!(dbg!(&mapping).len(), expected_mapping_len);
-        for (reduced_range, markdown_range) in mapping.iter() {
-            assert_eq!(
-                dbg!(crate::util::sub_chars(&plain, reduced_range.clone())),
-                dbg!(crate::util::sub_chars(&input, markdown_range.clone()))
-            );
-        }
-    }
-
-    #[test]
-    fn emoji() {
-        cmark_reduction_test(
-            r#"
-Abcd
-
----
-
-e🌡🍁
-
----
-
-fgh"#,
-            r#"Abcd
-
-
-e🌡🍁
-
-
-fgh"#,
-            3,
-        );
-    }
-
-    #[test]
-    fn link_footnote() {
-        cmark_reduction_test(
-            r#"footnote [^linktxt]. Which one?
-
-[linktxt]: ../../reference/index.html"#,
-            r#"footnote linktxt. Which one?"#,
-            3,
-        );
-    }
-
-    #[test]
-    fn link_inline() {
-        cmark_reduction_test(
-            r#" prefix [I'm an inline-style link](https://duckduckgo.com) postfix"#,
-            r#"prefix I'm an inline-style link postfix"#,
-            3,
-        );
-    }
-    #[test]
-    fn link_auto() {
-        cmark_reduction_test(
-            r#" prefix <http://foo.bar/baz> postfix"#,
-            r#"prefix  postfix"#,
-            2,
-        );
-        cmark_reduction_test(r#" <http://foo.bar/baz>"#, r#""#, 0);
-    }
-
-    #[test]
-    fn link_email() {
-        cmark_reduction_test(
-            r#" prefix <loe@example.com> postfix"#,
-            r#"prefix  postfix"#,
-            2,
-        );
-    }
-
-    #[test]
-    fn link_reference() {
-        cmark_reduction_test(
-            r#"[classy reference link][the reference str]"#,
-            r#"classy reference link"#,
-            1,
-        );
-    }
-
-    #[test]
-    fn link_collapsed_ref() {
-        cmark_reduction_test(
-            r#"[collapsed reference link][]"#,
-            r#"collapsed reference link"#,
-            1,
-        );
-    }
-
-    #[test]
-    fn link_shortcut_ref() {
-        cmark_reduction_test(
-            r#"[shortcut reference link]"#,
-            r#"shortcut reference link"#,
-            1,
-        );
-    }
-    // Nested links as well as nested code blocks are
-    // impossible according to the common mark spec.
-
-    #[test]
-    fn list_nested() {
-        cmark_reduction_test(
-            r#"
-* [x] a
-* [ ] b
-  * [ ] c
-  * [x] d
-"#,
-            r#"
-a
-b
-c
-d"#,
-            4,
-        );
-    }
-
-    #[test]
-    fn table_ignore() {
-        // TODO FIXME it would be better to transform this into
-        // one line per cell and test each cell.
-        // TODO very most likely will cause issues with grammar checks
-        // so eventually this will have to become checker specific code
-        // or handle a list of mute tags to simply ignore.
-        cmark_reduction_test(
-            r#"
-00
-
-|a|b|c
-|-|-|-
-|p|q|r
-
-ff
-"#,
-            r#"00
-
-
-ff"#,
-            2,
-        );
     }
 }
